+++ public/css/app.css
@@ -163,3 +163,4 @@ a {
 .icon_md {
     font-size: 2em !important;
 }
+
