+++ public/mix-manifest.json
@@ -1,4 +1,5 @@
 {
     "/js/app.js": "/js/app.js",
-    "/css/app.css": "/css/app.css"
+    "/css/app.css": "/css/app.css",
+    "/css/meetup.css": "/css/meetup.css"
 }
