+++ resources/css/app.css
@@ -1,5 +1,4 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
-
+@import url(https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap);
 /* Variables */
 :root {
 
@@ -23,6 +22,9 @@
 
     --link: #007bff;
     --selection: #90c6ff;
+
+    --border: #606770;
+    --pink: #f42da8;
 
     font-family: "Poppins", sans-serif;
 }
@@ -79,6 +81,7 @@ body {
     display: flex;
     flex-direction: column;
     gap: 1em;
+    width: -moz-fit-content;
     width: fit-content;
     padding: 1em;
     border-right: var(--neutral-400) solid 1px;
