+++ webpack.mix.js
@@ -12,6 +12,9 @@ const mix = require('laravel-mix');
  */
 
 mix.js('resources/js/app.js', 'public/js')
+    .postCss('resources/css/meetup.css', 'public/css')
     .postCss('resources/css/app.css', 'public/css', [
         //
     ]);
+
+
